feat: move proposals borsh schema
